initial refactor of custom HTML block to modernize
